News Posts: Authorization

+++ Source/Website/HoaLacLaptopShop/Controllers/NewsPostsController.cs
@@ -7,6 +7,8 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using HoaLacLaptopShop.Models;
 using HoaLacLaptopShop.ViewModels;
+using Microsoft.AspNetCore.Authorization;
+using HoaLacLaptopShop.Helpers;
 
 namespace HoaLacLaptopShop.Controllers
 {
@@ -19,7 +21,6 @@ namespace HoaLacLaptopShop.Controllers
             _context = context;
         }
 
-        // GET: NewsPosts
         public async Task<IActionResult> Index(NewsPostIndexArgs? args = null)
         {
             var news = _context.NewsPosts.Include(n => n.Author).OrderByDescending(x => x.Time).AsQueryable();
@@ -37,7 +38,6 @@ namespace HoaLacLaptopShop.Controllers
             });
         }
 
-        // GET: NewsPosts/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -56,32 +56,32 @@ namespace HoaLacLaptopShop.Controllers
             return View(newsPost);
         }
 
-        // GET: NewsPosts/Create
+        [Authorize(Roles = "Sales")]
         public IActionResult Create()
         {
-            ViewData["AuthorId"] = new SelectList(_context.Users, "ID", "ID");
             return View();
         }
 
-        // POST: NewsPosts/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,Title,Content,AuthorId")] NewsPost newsPost)
+        [Authorize(Roles = "Sales")]
+        public async Task<IActionResult> Create([Bind("ID,Title,Content")] NewsPost newsPost)
         {
+            ModelState.Remove(nameof(NewsPost.AuthorId));
+
             if (ModelState.IsValid)
             {
                 newsPost.Time = DateTime.Now;
+                newsPost.AuthorId = HttpContext.GetCurrentUser()!.ID;
+
                 _context.Add(newsPost);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["AuthorId"] = new SelectList(_context.Users, "ID", "ID", newsPost.AuthorId);
             return View(newsPost);
         }
 
-        // GET: NewsPosts/Edit/5
+        [Authorize(Roles = "Sales")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -94,33 +94,42 @@ namespace HoaLacLaptopShop.Controllers
             {
                 return NotFound();
             }
-            ViewData["AuthorId"] = new SelectList(_context.Users, "ID", "ID", newsPost.AuthorId);
+            if (!HttpContext.GetCurrentUser()!.IsAdmin && newsPost.AuthorId != HttpContext.GetCurrentUserID())
+            {
+                return Unauthorized();
+            }
+
             return View(newsPost);
         }
 
-        // POST: NewsPosts/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Content,AuthorId")] NewsPost newsPost)
+        [Authorize(Roles = "Sales")]
+        public async Task<IActionResult> Edit(int? id, [Bind("ID,Title,Content")] NewsPost newsPost)
         {
-            if (id != newsPost.ID)
+            if (id != newsPost.ID) return NotFound();
+            var targetPost = await _context.NewsPosts.FindAsync(id);
+            if (targetPost is null) return NotFound();
+
+            if (!HttpContext.GetCurrentUser()!.IsAdmin && targetPost.AuthorId != HttpContext.GetCurrentUserID())
             {
-                return NotFound();
+                return Unauthorized();
             }
 
             if (ModelState.IsValid)
             {
                 try
                 {
-                    newsPost.Time = DateTime.Now;
-                    _context.Update(newsPost);
+                    targetPost.Title = newsPost.Title;
+                    targetPost.Content = newsPost.Content;
+                    targetPost.Time = DateTime.Now;
+
+                    _context.Update(targetPost);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!NewsPostExists(newsPost.ID))
+                    if (!NewsPostExists(targetPost.ID))
                     {
                         return NotFound();
                     }
@@ -129,13 +138,12 @@ namespace HoaLacLaptopShop.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Details), new { id = newsPost.ID });
+                return RedirectToAction(nameof(Details));
             }
-            ViewData["AuthorId"] = new SelectList(_context.Users, "ID", "ID", newsPost.AuthorId);
             return View(newsPost);
         }
 
-        // GET: NewsPosts/Delete/5
+        [Authorize(Roles = "Sales")]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -143,25 +151,31 @@ namespace HoaLacLaptopShop.Controllers
                 return NotFound();
             }
 
-            var newsPost = await _context.NewsPosts
-                .Include(n => n.Author)
-                .FirstOrDefaultAsync(m => m.ID == id);
+            var newsPost = await _context.NewsPosts.Include(n => n.Author).FirstOrDefaultAsync(m => m.ID == id);
             if (newsPost == null)
             {
                 return NotFound();
+            }
+            if (!HttpContext.GetCurrentUser()!.IsAdmin && newsPost.AuthorId != HttpContext.GetCurrentUserID())
+            {
+                return Unauthorized();
             }
 
             return View(newsPost);
         }
 
-        // POST: NewsPosts/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Sales")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var newsPost = await _context.NewsPosts.FindAsync(id);
             if (newsPost != null)
             {
+                if (!HttpContext.GetCurrentUser()!.IsAdmin && newsPost.AuthorId != HttpContext.GetCurrentUserID())
+                {
+                    return Unauthorized();
+                }
                 _context.NewsPosts.Remove(newsPost);
             }
